requirement.txt
absl-py==0.8.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.5
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bio==0.1.0
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
Boruta==0.3
boto==2.49.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20181111
distlib==0.3.0
distributed==1.26.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.41.0
django-cors-headers==3.2.1
django-grappelli==2.14.1
django-import-export==2.0.2
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-reverse-admin==2.8.3
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.14
editdistance==0.5.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.10
Flask==1.0.2
future==0.17.1
gast==0.2.2
genepattern-python==1.4.3
gevent==1.4.0
glob2==0.6
gluonnlp==0.9.1
google-pasta==0.1.7
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.24.1
h5py==2.9.0
HeapDict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.5
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10
joblib==0.14.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.3
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
leven==1.0.4
libarchive-c==2.8
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.16.0
mock==3.0.5
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mxboard==0.1.0
mxnet==1.6.0
mysqlclient==1.4.4
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.3
opt-einsum==3.1.0
orca==1.5.3
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pdf2image==1.12.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.10
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.10.0
psutil==5.6.1
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyQt5==5.12
PyQt5-sip==4.19.19
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytesseract==0.3.4
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-Levenshtein-wheels==0.13.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.9
PyWavelets==1.0.2
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
regex==2020.4.4
reportlab==3.5.34
requests==2.18.4
requests-oauthlib==1.3.0
rope==0.12.0
ruamel.yaml==0.15.46
sacremoses==0.0.41
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.3
tables==3.5.1
tablib==1.1.0
tblib==1.3.2
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.7.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
yarg==0.1.9
zict==0.1.4
zipp==0.3.3
